renamed  package to galmock
