+++ galmock/__init__.py
