+++ galmock/core/bitmask.py
