+++ galmock/core/parallel.py
@@ -12,7 +12,7 @@ import numpy as np
 from mmaptable.column import MmapColumn
 from mmaptable.table import MmapTable
 
-from mock_processing.core.utils import ProgressBar
+from galmock.core.utils import ProgressBar
 
 
 class Schedule:
