+++ galmock/core/utils.py
