+++ galmock/matching.py
@@ -7,11 +7,11 @@ from scipy.spatial import cKDTree
 
 from mmaptable import MmapTable
 
-from mock_processing.core.config import (Parameter, ParameterCollection,
+from galmock.core.config import (Parameter, ParameterCollection,
                                          ParameterGroup, ParameterListing,
                                          Parser)
-from mock_processing.core.parallel import Schedule
-from mock_processing.core.utils import expand_path
+from galmock.core.parallel import Schedule
+from galmock.core.utils import expand_path
 
 
 class DistributionEstimator(object):
