+++ galmock/samples/Flagship.py
@@ -1,0 +1,2 @@
+from galmock.samples import reference
+from galmock.samples.base import SampleManager
