+++ galmock/samples/data/Sparse24mag.json
