+++ galmock/samples/data/VVDSf02_z_bright_success
