+++ galmock/samples/data/VVDSf02_z_deep_success
