+++ galmock/samples/data/zCOSMOS.json
