+++ galmock/samples/data/zCOSMOS_I_sampling
