+++ galmock/samples/data/zCOSMOS_success
