+++ galmock/samples/data/zCOSMOS_z_sampling
