--- mock_processing/samples/Flagship.py
@@ -1,2 +1,0 @@
-from mock_processing.samples import reference
-from mock_processing.samples.base import SampleManager
