+++ scripts/Flagship_flux_to_magnitudes
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import argparse
 
-from mock_processing.jobs import flux_to_magnitudes
+from galmock.jobs import flux_to_magnitudes
 
 
 parser = argparse.ArgumentParser(
