+++ scripts/MICE2_evolution_correction
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import argparse
 
-from mock_processing.jobs import evolution_correction
+from galmock.jobs import evolution_correction
 
 
 parser = argparse.ArgumentParser(
