+++ scripts/mocks_convert_input
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 import argparse
 
-from mock_processing.core.readwrite import SUPPORTED_READERS
-from mock_processing.jobs import convert_input
+from galmock.core.readwrite import SUPPORTED_READERS
+from galmock.jobs import convert_input
 
 
 parser = argparse.ArgumentParser(
