+++ scripts/mocks_datastore_info
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import argparse
 
-from mock_processing.jobs import datastore_info
+from galmock.jobs import datastore_info
 
 
 parser = argparse.ArgumentParser(
