+++ scripts/mocks_datastore_query
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 import argparse
 
-from mock_processing.core.readwrite import SUPPORTED_WRITERS
-from mock_processing.jobs import datastore_query
+from galmock.core.readwrite import SUPPORTED_WRITERS
+from galmock.jobs import datastore_query
 
 
 parser = argparse.ArgumentParser(
