+++ scripts/mocks_datastore_verify
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import argparse
 
-from mock_processing.jobs import datastore_verify
+from galmock.jobs import datastore_verify
 
 
 parser = argparse.ArgumentParser(
