+++ scripts/mocks_init_pipeline
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 import argparse
 
-from mock_processing.core.config import TableParser
-from mock_processing.core.readwrite import SUPPORTED_READERS
-from mock_processing.jobs import init_pipeline
+from galmock.core.config import TableParser
+from galmock.core.readwrite import SUPPORTED_READERS
+from galmock.jobs import init_pipeline
 
 
 parser = argparse.ArgumentParser(
