+++ scripts/mocks_magnification
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import argparse
 
-from mock_processing.jobs import magnification
+from galmock.jobs import magnification
 
 
 parser = argparse.ArgumentParser(
