+++ scripts/mocks_select_sample
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 import argparse
 
-from mock_processing.core.readwrite import SUPPORTED_WRITERS
-from mock_processing.jobs import select_sample
-from mock_processing.samples import DumpConfig, SampleManager
+from galmock.core.readwrite import SUPPORTED_WRITERS
+from galmock.jobs import select_sample
+from galmock.samples import DumpConfig, SampleManager
 
 
 parser = argparse.ArgumentParser(
